switch to using conf.PublicFQDN for tink_host to resolve phone-home issues
Previously the tink_host function pointed to conf.PublicIPv4 which doesn't
work well for virtual host based http servers, as the Host header being
passed was the IP rather than a domain name.

This is resolved by using conf.PublicFQDN which, if the PUBLIC_FQDN
environment variable is not set, falls back to conf.PublicIPv4 but allows
the administrator to set the PUBLIC_FQDN environment variable to override
this with a domain name. ex: boots.example.com