include syslog host in osie kernel environments
This allows for a separate host to be defined as the syslog server.
The default action is to fallback to the `PublicIPv4` which is what
is currently used as the DHCP Log Server option value.

You can now define both `PUBLIC_SYSLOG_IP` and `PUBLIC_SYSLOG_FQDN`.
The former is needed to change the dhcp log server. The latter is
passed into the ipxe script.

If the fqdn is not defined, it falls back to the value of the ip.
If the ip is not defined it falls back to the public ip.

Signed-off-by: Mike Mason <mason@packet.com>