Replace golangci-lint from the action with Nix
The action keeps spewing un-tar errors and we already have
golangci-lint available via nix (:surprise:) so lets use it.

I hope this paves the way for using tools from nix instead of actions
(and w/e versions those bring in).

Not sure why the setup-go@v2 action is setting GOROOT, but it's messing
with the Nix goloangci-lint run as it complains about go tool version
mismatches.

Signed-off-by: Manuel Mendez <mmendez@equinix.com>