tftp: Flip early return in ReadFile
Previously we were returning early if err==nil when looking up a job, but when
we extracted the HasActiveWorkflow check we ended up changing behavior.

job.CreateFromIP would not have returned err==nil unless we were able to fetch
workflows and there was in fact an active workflow. So we now need to flip it
to handle error cases instead of happy case so that we can check each possible
error case.