workflows: Push to Quay from an entirely differnt workflow
This is the key piece that allows us to push images for PR builds.
This workflow is run from the main branch's context/code and not the PR's
and so has access to secrets and can thus log in to push to Quay.

This does break the normal workflow _flow_ that we're used to but it seems
better to avoid duplicating a lot of yaml just to maintain the current flow
of pushing to quay from master via ci.yaml.

Signed-off-by: Manuel Mendez <mmendez@equinix.com>