job: test: Minor changes to TestSetPXEFilename cases
More context in name is just for human context. Added AllowPXE
explicitly because the next commit is going to change some of the logic
and I don't want to change the code and the tests at the same time.
All of these cases that return a valid filename either don't care
about allowPXE or wanted it set anyway.

Signed-off-by: Manuel Mendez <mmendez@equinix.com>