Drop Packet from iPXE version checks and use
Just getting rid of a Packetism.

We override the iPXE version that iPXE uses for itself and sends to the dhcp
server. This is how we distinguish an iPXE burned into a nic by a vendor vs the
one we build and send via tftp. There are better options here, but not as
convenient to apply. We can define the IPXE_VERSION via an env var when building
ipxe but other options would require a patch to one or more locations in the
iPXE code. This version number is no longer incrementing in upstream iPXE so is
low-risk to modify.

Signed-off-by: Manuel Mendez <mmendez@equinix.com>