dhcp: Move setting PXEClient ClassID to dhcp.SetupPXE
Section 2.2.2 of the Intel PXE Spec states:

> The purpose of the timeouts is to ensure the PXE client gives precedence to servers supplying
> “PXEClient” specific configuration tags. The PXE boot ROM must function as a normal DHCP boot
> ROM in the absence of a PXE specific response. However, the PXE boot ROM must wait for
> specified times to see if a PXE response is available before using a non-PXE configuration.

Which sort of explains why PXE clients that are not *our* iPXE build will still
take the `filename` offered up even though the ClassID was not set for
PXEClient when job/dhcp.go:setPXEFilename:`if !isOuriPXE` is the branch taken.

SetupPXE seems like the best place for this setting any way as we setup the PXE
response packet without need for job information.

Signed-off-by: Manuel Mendez <mmendez@equinix.com>