Fix lost Make idempotency with ipxe builds:
With the addition of bin/ipxe.lkrn the bin/undionly.kpxe
build directory was being overwritten by bin/ipxe.lkrn.
This caused Make to always build both of these targets
and not provide the idempotency we had prior. Moving
bin/ipxe.lkrn to bin-test/ipxe.lkrn is a solution to
regain the Make idempotency.

Signed-off-by: Jacob Weinstock <jakobweinstock@gmail.com>