Add OSIE secrets required by Hollow
In order for OSIE to write data to Hollow, we need to use these secrets
to request a temporary auth token.

We recognize that Equinix Metal specific code needs to be extracted from
Boots, and are looking into general solutions that will allow users to
specify secrets or custom kernel boot parameters.

Signed-off-by: Scott Garman <sgarman@equinix.com>