start adding spans
At this point a DHCP request that matches on config will emit some spans
and events. This is a first pass and will need some tuning to look good
but it's all working and emitting data.

The big change in the code is adding context.Context passing to
functions that didn't have it before. I chose this route rather than
directly passing trace.Span because it's a well-supported path and the
context may be useful for other things like request cancelation later.

We might be able to drop the context.Context from the Job struct if
adding context passing to various functions does the trick.

Signed-off-by: Amy Tobey <atobey@equinix.com>