use github.com/tobert/otel-init-go to initialize otel
This moves the boilerplate opentelemetry setup code into a standalone
module that can be reused across services. It's mostly the same code
that was in opentelemetry.go except that it is only configurable via
environment variables.

https://github.com/tobert/otel-init-go

We might want to move this into a trusted org before running with it in
my personal GitHub. Once I get it shaken out in a few other projects
I'll move it to a PR model with releases and stuff so it's a stable
target so maybe it's fine where it is. Happy to discuss.

Signed-off-by: Amy Tobey <atobey@equinix.com>