Get rid of bouk/monkey dependency
bouk/monkey's license grants no rights of use and explicitly states that it
should not be used. The license text is very clear:

```
Copyright Bouke van der Bijl

I do not give anyone permissions to use this tool for any purpose. Don't use it.

I’m not interested in changing this license. Please don’t ask.
```

So this commit brings boots back into a compliance with its dependencies.

I've gotten rid of the tests that used the dependency entirely, for now. I'll open
up an issue to re-introduce such tests and will try to get to it soon.

Fixes #196

Signed-off-by: Manuel Mendez <mmendez@equinix.com>