return context from job.CreateFromDHCP
In the previous commits, I added the ability for traceparent to be
communicated in-band through hardware records. Once that arrives, the
job module sees it and updates the context with that traceparent for
future span. This does no good until that context is returned up the
call stack to cmd/dhcp.go where it can use that ctx going forward for
the rest of the DHCP transaction, including sending the traceparent in a
DHCP option.

Signed-off-by: Amy Tobey <atobey@equinix.com>