modify and propagate context with upstream tp
Modifies the job.setup code to swap out the context on its way through
so that the TraceContext points at the traceparent we got from the
hardware record.

Adds context passing to relevant methods.

Modifies the http ipxe file handler to req.Clone(ctx) to swap that new
context in so internal spans end up linked to the upstream traces.

Signed-off-by: Amy Tobey <atobey@equinix.com>