add a second trace to the tftp server
Adds another trace and link to the tftp server code. I'm not super happy
with the double trace but it seems to make sense.

If the link and everything works out maybe we can drop the filename
hack?

I wasn't able to validate the links coming through because the otel
collector and otel-cli don't show them.

Signed-off-by: Amy Tobey <atobey@equinix.com>