ci: Allow pushing to image registry from forks
This will enable forks to use our ci configuration to push their own images to
their own image registry if they set the necessary secrets. This is done in the
"push" context which isn't that big of a deal really since the user could just
as easily push from their development machine, but this way its done
automatically by the computers in github's cloud. Automated > manual, always.
If the branch is rebased on top of tinkerbell's main then the build/push would
be equivalent to the push that would happen post merge.

Signed-off-by: Manuel Mendez <mmendez@equinix.com>