Always use ipxe.Script as a pointer
There's only one actual case where returning the Script was nice (the
`Set("iface",...).Or("shell")`) but its not worth keeping around. Its confusing
seeing the script being modified in place within a function body but having
to copy and return within and after passing to a function. Now we know the
script is always modified in place.

Signed-off-by: Manuel Mendez <github@i.m.mmlb.dev>