ci: Misc updates to ci.yaml
Change from self-hosted to GitHub managed `ubuntu-latest` because
install-nix-action@v17 fails with an error due to missing python. We don't
have a huge need for the self hosted runner since we no longer build ipxe
and life is just easier this way.

Bump cachix/install-nix-action is bumped because v17 contains 018abf9.

Pin actions/setup-go to 75726802 until a new release comes out, this commit
is when my PR was merged that stops adding GOROOT to the environment which
causes issues with the go used via nix-shell.

Signed-off-by: Manuel Mendez <github@i.m.mmlb.dev>