Add CLI flag for overriding OSIE path
If an OSIE download path URL is specified via this fkag, OSIE_PATH and
MIRROR_HOST variables that are used to build OSIE URL will be ignored.
Hardcoded appending of 'misc/osie' and 'current' directories to the path
will be bypassed.

With the ongoing refactoring efforts, once all this hardcoded appending
of paths is cleaned up, this flag can be removed/modified

Signed-off-by: Pooja Trivedi <tripooja@amazon.com>