Allow extra kernel args in iPXE for OSIE, set via cli flag:
This change is targeted at operators of Hook. This allows
operators to set any necessary and needed
kernel commandline args to the default installer; OSIE.
This makes custom kernel args a first class option instead
of the current model to overload the "facility" field in
the backend data. Kernel args set with this CLI flag will
be global to all iPXE boots.

Signed-off-by: Jacob Weinstock <jakobweinstock@gmail.com>