properly set overwritten password
The provided on an instance is typically a generated password.
For VMWare an alternative password hash can be provided via the
instance's customdata field.

Previously the method for utilizing that override required manipulating the shadow file.
This method is no longer supported by vmware, as the shadow file is
generated based off a database value.

The correct way to set the root password is through the kickstart file.
We already were setting the generated password, so we could simply add
the necessary logic to check for an overwritten password instead.

Signed-off-by: Mike Mason <mimason@equinix.com>