Split docker build and push into different steps
Doing both in one step is a little confusing, it can lead to thinking a
PR's build will be pushed. We need to go in and expand the build logs to
see if the build was actually pushed. By splitting it like this it makes it
a little more obvious that Pushing is going to be skipped.

Signed-off-by: Manuel Mendez <github@i.m.mmlb.dev>