Refactor auto.ipxe and custom ipxe scripts:
This refactor seeks to make it more clear for developers what the
auto.ipxe script will be. Same for what was previously called
custom_ipxe. No longer is it needed to specify the operating
system to be "custom_ipxe". If
hardware.spec.interfaces[0].netboot.ipxe.url or
hardware.spec.interfaces[0].netboot.ipxe.contents is specified then
it is used instead of the default auto.ipxe for Hook.

Signed-off-by: Jacob Weinstock <jakobweinstock@gmail.com>