Remove trusted proxies around auto.ipxe:
Move from using IP auth via the source IP address
to using the MAC address in the URL in order to determine
the Hardware to lookup when serving the auto.ipxe.
This opens up the auto.ipxe script to be able to be served
across network address translation (NAT) layers and simplifies
the deployment.

Signed-off-by: Jacob Weinstock <jakobweinstock@gmail.com>