Move DHCP library back into Smee:
The DHCP library was created in order to
provide a mechanism for the DHCP functionality
to be written in a very clean and composible way.
This was not possible with the previous state of Boots
and the personnel involved at that time.

The current state of Smee is ready for this functionality
to be return and placed in a very clean and composible way
into the code base.

Moving this functionality back in will improve the development
process/experience. It is important to note that we do take on
the trade off of not having the combined commit history of the library
within Smee. The library GitHub repo will still contain history.

Signed-off-by: Jacob Weinstock <jakobweinstock@gmail.com>