Add trusted proxies back as optional:
There is an existing use-case where a user doesn't use Smee
for DHCP and interpolating the mac address into the URL for
the auto.ipxe script is either not possible or very difficult
in their DHCP server.
For this reason, we will allow using trusted proxies (XFF).
We will also fall back to checking for the source IP in a request
if the MAC address is not found in the URL path.

Signed-off-by: Jacob Weinstock <jakobweinstock@gmail.com>