Refactor the patching implementation:
The previous implementation only worked for
clients that did partial content requests.
This refactor allows all requests, whether
partial content or not to have patching occur.

This is particularly useful when using the
redfish emulator from sushy-tools. This emulator
doesn't not request the ISO as partial content
but requests the whole ISO like a curl or wget client
might.

This implementation is very memory safe and doesn't
require content length checking of any kind.

The Go stdlib httputil.NewSingleHostReverseProxy file
and supporting files were copied in. This is because
reimplementing all the logic in func (p *ReverseProxy) ServeHTTP
of the httputil package is a lot and made more sense to
copy it in and modify it to be able to patch when writing
the body to the client.

Signed-off-by: Jacob Weinstock <jakobweinstock@gmail.com>